[WIP] "Products We Love" section for Landing Page of Reaction Swag Shop
This PR resolves issues #13 and #19. As it's a work-in-progress, not
every aspect of the user stories is addressed yet.

(feat): As an admin, allow to specifify a new text property `featuredProductLabel`
for products. All products that have specified this property will be
displayed in section "Products We Love". The propertie's value will be
displayed as turquoise label in the upper left corner of the product
image.

Not functional yet:
Acceptance criteria "Only 1 featured product is shown" for mobile
devices is not met. Right now, all featured products are shown (i.e. up to
productScrollLimit = 24 items)